Finished 01-margin-padding
The task went over how to add margin and padding
to divs to acheive a desired outcome. The task
covered using pixels as values for padding and
margin values. The challenge of this task was to
understand how to use the value of "auto" for margin
to align items. The desired outcome was met properly.