Finished 02-margin-and-padding
This task goes over how to manipulate divs to look
a certain way. It was an approach to a real world
scenario where you're just given divs without any
values of padding and margins and your job is to
manipulate it to make it look a certain way. The
desired outcome was reached however the challenges
that I had to tackle was to use the auto value
properly to center items within a div as well as
center the entire div on the webpage. Another tackle
was to use block display to position items in their
own line.