Finished 02-flex-header
This task consisted of manipulating the navigation
header of a webpage to match the desired outcome.
This task covered how to use the flex-box syntax
to align items, justify content as well as regular
CSS styles like padding, margin, gap. The CSS file
has 2 solutions, the commented ones are the 2nd
set of solutions.