Finished 04-flex-information
Was able to acheive the desired outcome. the task
went over how to utilize justify-content to center
elements horizontally in a webpage and use flexbox.
The challenge of this task was to understand that
max-width had to be used instead of width to not allow
the width to be altered past 200px.